requirements.txt
flask 
opencv-python 
pillow 
gunicorn 
Jinja2 
h5py 
itsdangerous 
MarkupSafe 
Werkzeug